账单模块

+++ src/main/java/com/zhiji/smartbook/module/bill/vo/BillGroupByDateVO.java
@@ -2,12 +2,13 @@ package com.zhiji.smartbook.module.bill.vo;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
 public class BillGroupByDateVO {
     private String date;
-    private Double expenseTotal;
-    private Double incomeTotal;
+    private BigDecimal expenseTotal;
+    private BigDecimal incomeTotal;
     private List<BillListItemVO> items;
 }
